+++ changelog.md
@@ -6,6 +6,7 @@
 This version is not released yet and is under active development.
 ```
 
+- Add a simple `--telemetry`/`--no-telemetry` option flag which respects the `DO_NOT_TRACK` environment variable.
 - Fix rendering of GitHub-Flavored Markdown tables in canonical format.
 
 ## {gh}`3.10.0 (2023-04-04) <compare/v3.9.0...v3.10.0>`